Changed long long int to int64_t for index_type

+++ src/multi_index_array/MappedSparseLattice.h
@@ -59,7 +59,7 @@ struct data_type<MappedSparseLattice<T> >
 template<typename T>
 struct index_type<MappedSparseLattice<T> >
 {
-    typedef long long type;
+    typedef int64_t type;
 };
 
 template<typename T>
